Moved query functions from api.py to model.py
